feat: add `post-guide` command
