+++ src/lib/i18n/LanguageKeys/Commands/All.ts
@@ -1,3 +1,4 @@
 export * as Config from '#lib/i18n/LanguageKeys/Commands/Config';
+export * as PostGuide from '#lib/i18n/LanguageKeys/Commands/PostGuide';
 export * as Resolve from '#lib/i18n/LanguageKeys/Commands/Resolve';
 export * as Suggest from '#lib/i18n/LanguageKeys/Commands/Suggest';
