+++ src/lib/i18n/LanguageKeys/InteractionHandlers/All.ts
@@ -1,2 +1,3 @@
 export * as Suggestions from '#lib/i18n/LanguageKeys/InteractionHandlers/Suggestions';
+export * as SuggestionsCreate from '#lib/i18n/LanguageKeys/InteractionHandlers/SuggestionsCreate';
 export * as SuggestionsModals from '#lib/i18n/LanguageKeys/InteractionHandlers/SuggestionsModals';
