+++ src/lib/i18n/LanguageKeys/InteractionHandlers/SuggestionsCreate.ts
@@ -1,0 +1,4 @@
+import { T } from '@skyra/http-framework-i18n';
+
+export const ModalTitle = T('interaction-handlers/suggestions-create:modalTitle');
+export const ModalLabel = T('interaction-handlers/suggestions-create:modalLabel');
