+++ src/locales/en-US/commands/post-guide.json
@@ -1,0 +1,11 @@
+{
+	"name": "post-guide",
+	"description": "Posts a guide on how to publish new suggestions",
+	"optionsHideName": "hide",
+	"optionsHideDescription": "Whether or not the guide should be hidden",
+	"optionsTargetName": "target",
+	"optionsTargetDescription": "The user to mention in the message, if any",
+	"message": "To submit a new suggestion, simply use the {{command}} command with {{bot}}, or click the \"Submit a Suggestion\" button attached.\n\nYour message will be posted at {{channel}}.",
+	"buttonSubmit": "Submit a Suggestion",
+	"buttonSupport": "Support Server"
+}
