+++ src/locales/en-US/interaction-handlers/suggestions-create.json
@@ -1,0 +1,4 @@
+{
+	"modalTitle": "Submit a Suggestion",
+	"modalLabel": "Content"
+}
